Moved samples to ImageSources. Added ImageSources gallery.
