+++ samples/CommunityToolkit.Maui.Sample/MauiProgram.cs
@@ -4,11 +4,13 @@ using CommunityToolkit.Maui.Sample.Pages.Alerts;
 using CommunityToolkit.Maui.Sample.Pages.Behaviors;
 using CommunityToolkit.Maui.Sample.Pages.Converters;
 using CommunityToolkit.Maui.Sample.Pages.Extensions;
+using CommunityToolkit.Maui.Sample.Pages.ImageSources;
 using CommunityToolkit.Maui.Sample.Pages.Layouts;
 using CommunityToolkit.Maui.Sample.Pages.Views;
 using CommunityToolkit.Maui.Sample.ViewModels.Alerts;
 using CommunityToolkit.Maui.Sample.ViewModels.Behaviors;
 using CommunityToolkit.Maui.Sample.ViewModels.Converters;
+using CommunityToolkit.Maui.Sample.ViewModels.ImageSources;
 using CommunityToolkit.Maui.Sample.ViewModels.Layouts;
 using CommunityToolkit.Maui.Sample.ViewModels.Views;
 using CommunityToolkit.Maui.Sample.ViewModels.Views.AvatarView;
@@ -42,6 +44,7 @@ public static class MauiProgram
 		services.AddTransient<BehaviorsGalleryPage, BehaviorsGalleryViewModel>();
 		services.AddTransient<ConvertersGalleryPage, ConvertersGalleryViewModel>();
 		services.AddTransient<ExtensionsGalleryPage, ExtensionsGalleryViewModel>();
+		services.AddTransient<ImageSourcesGalleryPage, ImageSourcesGalleryViewModel>();
 		services.AddTransient<LayoutsGalleryPage, LayoutsGalleryViewModel>();
 		services.AddTransient<ViewsGalleryPage, ViewsGalleryViewModel>();
 
@@ -117,12 +120,14 @@ public static class MauiProgram
 		// Add Extensions Pages + ViewModels
 		services.AddTransient<ColorAnimationExtensionsPage, ColorAnimationExtensionsViewModel>();
 
+		// Add ImageSources pages + ViewModels
+		services.AddTransient<GravatarImageSourcePage, GravatarImageSourceViewModel>();
+
 		// Add Layouts Pages + ViewModels
 		services.AddTransient<UniformItemsLayoutPage, UniformItemsLayoutViewModel>();
 
 		// Add Views Pages + ViewModels
 		services.AddTransient<DrawingViewPage, DrawingViewViewModel>();
-		services.AddTransient<GravatarImageSourcePage, GravatarImageSourceViewModel>();
 		services.AddTransient<MultiplePopupPage, MultiplePopupViewModel>();
 		services.AddTransient<PopupAnchorPage, PopupAnchorViewModel>();
 		services.AddTransient<PopupPositionPage, PopupPositionViewModel>();
